Minimizing logic in amp-ad-network-base.
